Rename fileUpload.js to fileHandler.js
